# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Deckwatch API
  description: |
    Kubernetes control-plane and OCI registry surface exposed by Deckwatch.

    - `/api/*` — application, deployment, namespace, and cluster management
    - `/v2/*`  — OCI Distribution Spec v1.1 (only when the embedded registry is enabled)
    - `/metrics` — Prometheus exposition
  version: "0.1.0"
  contact:
    name: Deckwatch
    url: https://github.com/kore/deckwatch
servers:
  - url: /
    description: Same-origin (served by deckwatch)
...
paths:
...
  /api/namespaces:
    get:
      tags: [Namespaces]
      summary: List namespaces visible to deckwatch
      responses:
        "200":
          description: Namespaces
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NamespaceListResponse" }
    post:
      tags: [Namespaces]
      summary: Create a namespace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateNamespaceRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateNamespaceResponse" }

  # -------------------------------------------------------------- deployments
  /api/namespaces/{ns}/deployments:
    parameters:
      - $ref: "#/components/parameters/Namespace"
    get:
      tags: [Deployments]
      summary: List deployments in a namespace
      parameters:
        - in: query
          name: label_selector
          required: false
          schema: { type: string }
          description: Kubernetes label selector (e.g. `app=api`)
      responses:
        "200":
          description: Deployments
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeploymentListResponse" }
    post:
      tags: [Deployments]
      summary: Create a deployment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDeploymentRequest" }
      responses:
        "201": { description: Created }
  /api/namespaces/{ns}/deployments/{name}:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    get:
      tags: [Deployments]
      summary: Get a deployment with pods and ingresses
      responses:
        "200":
          description: Deployment detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeploymentDetailResponse" }
    put:
      tags: [Deployments]
      summary: Update a deployment (image, replicas, env, resources, probes)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateDeploymentRequest" }
      responses:
        "200": { description: Updated }
    delete:
      tags: [Deployments]
      summary: Delete a deployment
      responses:
        "204": { description: Deleted }
  /api/namespaces/{ns}/deployments/{name}/yaml:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    get:
      tags: [Deployments]
      summary: Get the raw Deployment manifest as YAML
      responses:
        "200":
          description: YAML manifest
          content:
            application/yaml:
              schema: { type: string }
    put:
      tags: [Deployments]
      summary: Replace the Deployment manifest from YAML
      requestBody:
        required: true
        content:
          application/yaml:
            schema: { type: string }
      responses:
        "200": { description: Updated }
  /api/namespaces/{ns}/deployments/{name}/restart:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    post:
      tags: [Deployments]
      summary: Rolling-restart a deployment
      responses:
        "202": { description: Restart triggered }
  /api/namespaces/{ns}/deployments/{name}/scale:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    post:
      tags: [Deployments]
      summary: Scale a deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [replicas]
              properties:
                replicas: { type: integer, minimum: 0 }
      responses:
        "200": { description: Scaled }
  /api/namespaces/{ns}/deployments/{name}/probes:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    patch:
      tags: [Deployments]
      summary: Patch liveness / readiness / startup probes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                liveness_probe:  { $ref: "#/components/schemas/ProbeInput" }
                readiness_probe: { $ref: "#/components/schemas/ProbeInput" }
                startup_probe:   { $ref: "#/components/schemas/ProbeInput" }
      responses:
        "200": { description: Updated }
  /api/namespaces/{ns}/deployments/{name}/containers:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    post:
      tags: [Deployments]
      summary: Add a sidecar container
      responses:
        "200": { description: Container added }
  /api/namespaces/{ns}/deployments/{name}/containers/{container_name}:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
      - in: path
        name: container_name
        required: true
        schema: { type: string }
    delete:
      tags: [Deployments]
      summary: Remove a sidecar container
      responses:
        "204": { description: Removed }
  /api/namespaces/{ns}/deployments/{name}/pods:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    get:
      tags: [Pods]
      summary: List pods belonging to a deployment
      responses:
        "200":
          description: Pods
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PodListResponse" }
  /api/namespaces/{ns}/deployments/{name}/history:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    get:
      tags: [Deployments]
      summary: List ReplicaSet revision history for a deployment
      responses: { "200": { description: Revisions } }
  /api/namespaces/{ns}/deployments/{name}/rollback:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    post:
      tags: [Deployments]
      summary: Rollback to a prior revision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                revision: { type: integer }
      responses:
        "202": { description: Rollback triggered }
  /api/namespaces/{ns}/deployments/validate:
    parameters:
      - $ref: "#/components/parameters/Namespace"
    post:
      tags: [Deployments]
      summary: Dry-run validate a deployment payload
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDeploymentRequest" }
      responses:
        "200": { description: Valid }
        "400": { description: Invalid }
  /api/namespaces/{ns}/deployments/{name}/clone:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    post:
      tags: [Deployments]
      summary: Clone a deployment (into same or another namespace)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_namespace: { type: string }
                target_name:      { type: string }
      responses:
        "201": { description: Cloned }

  # -------------------------------------------------------------- ingresses
  /api/namespaces/{ns}/ingresses:
    parameters:
      - $ref: "#/components/parameters/Namespace"
    get:
      tags: [Ingresses]
      summary: List ingresses
      responses: { "200": { description: OK } }
    post:
      tags: [Ingresses]
      summary: Create an ingress
      responses: { "201": { description: Created } }
  /api/namespaces/{ns}/ingresses/{name}:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    get:
      tags: [Ingresses]
      responses: { "200": { description: OK } }
    put:
      tags: [Ingresses]
      responses: { "200": { description: Updated } }
    delete:
      tags: [Ingresses]
      responses: { "204": { description: Deleted } }

  # -------------------------------------------------------------- gitops
  /api/namespaces/{ns}/deployments/{name}/gitops:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    get:
      tags: [GitOps]
      summary: Get GitOps config attached to a deployment
      responses:
        "200":
          description: GitOps status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GitOpsStatusResponse" }
    put:
      tags: [GitOps]
      summary: Attach or update GitOps config for a deployment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GitOpsConfigRequest" }
      responses: { "200": { description: Saved } }
    delete:
      tags: [GitOps]
      summary: Detach GitOps config
      responses: { "204": { description: Detached } }
  /api/namespaces/{ns}/deployments/{name}/gitops/trigger:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    post:
      tags: [GitOps]
      summary: Manually trigger a build (kaniko job)
      responses: { "202": { description: Build queued } }
  /api/namespaces/{ns}/deployments/{name}/gitops/builds:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Name"
    get:
      tags: [GitOps]
      summary: List past kaniko builds for this deployment
      responses: { "200": { description: OK } }

  # -------------------------------------------------------------- pods / logs / exec
  /api/namespaces/{ns}/pods/{pod_name}:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - in: path
        name: pod_name
        required: true
        schema: { type: string }
    get:
      tags: [Pods]
      summary: Get a pod
      responses:
        "200":
          description: Pod
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PodSummary" }
  /api/namespaces/{ns}/pods/{pod_name}/logs:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - in: path
        name: pod_name
        required: true
        schema: { type: string }
    get:
      tags: [Logs]
      summary: Stream pod logs (Server-Sent Events)
      parameters:
        - in: query
          name: container
          schema: { type: string }
        - in: query
          name: follow
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: text/event-stream
          content:
            text/event-stream:
              schema: { type: string }
  /api/namespaces/{ns}/pods/{pod_name}/logs/history:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - in: path
        name: pod_name
        required: true
        schema: { type: string }
    get:
      tags: [Logs]
      summary: Fetch a historical log slice (non-streaming)
      parameters:
        - in: query
          name: container
          schema: { type: string }
        - in: query
          name: tail_lines
          schema: { type: integer }
      responses: { "200": { description: Logs } }
...
  /api/namespaces/{ns}/cronjobs:
    parameters:
      - $ref: "#/components/parameters/Namespace"
    get:
      tags: [CronJobs]
      responses: { "200": { description: OK } }
...
components:
  parameters:
    Namespace:
      in: path
      name: ns
      required: true
      schema: { type: string }
      description: Kubernetes namespace
    Name:
      in: path
      name: name
      required: true
      schema: { type: string }
      description: Resource name
  schemas:
...
    NamespaceListResponse:
      type: object
      properties:
        namespaces:
          type: array
          items: { type: string }
    CreateNamespaceRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        labels:
          type: object
          additionalProperties: { type: string }
    CreateNamespaceResponse:
      type: object
      properties:
        name: { type: string }
        created_at: { type: string, format: date-time, nullable: true }
        labels:
          type: object
          additionalProperties: { type: string }
...
    DeploymentListResponse:
      type: object
      properties:
        deployments:
          type: array
          items: { $ref: "#/components/schemas/DeploymentSummary" }
...